5379: move empty list component
